+++ docs/source/changelog.rst
@@ -2,6 +2,11 @@
 Change Log
 **********
 
+**Pyro 4.53**
+
+- fixed some typos in docs and docstrings, improved daemon metadata doc.
+
+
 **Pyro 4.52**
 
 - fixed Python 3.6 compatibility issue in name server when using sqlite storage ("cannot VACUUM from within a transaction")